refactor(app-review): remove appReview store and move to app store
