--- src/appReview/selectors.ts
@@ -1,4 +1,0 @@
-import { RootState } from 'src/redux/reducers'
-
-export const lastInteractionTimestampSelector = (state: RootState) =>
-  state.appReview.lastInteractionTimestamp
