+++ src/redux/reducers.ts
@@ -4,7 +4,6 @@ import { Actions } from 'src/account/actions'
 import { reducer as account, State as AccountState } from 'src/account/reducer'
 import { reducer as alert, State as AlertState } from 'src/alert/reducer'
 import { appReducer as app, State as AppState } from 'src/app/reducers'
-import appReviewReducer, { State as AppReviewState } from 'src/appReview/slice'
 import superchargeReducer, { State as SuperchargeState } from 'src/consumerIncentives/slice'
 import dappsReducer, { State as DappsState } from 'src/dapps/slice'
 import { escrowReducer as escrow, State as EscrowState } from 'src/escrow/reducer'
@@ -54,7 +53,6 @@ const appReducer = combineReducers({
   fiatConnect: fiatConnectReducer,
   swap: swapReducer,
   positions: positionsReducer,
-  appReview: appReviewReducer,
 }) as (state: RootState | undefined, action: Action) => RootState
 
 const rootReducer = (state: RootState | undefined, action: Action): RootState => {
@@ -103,7 +101,6 @@ export interface RootState {
   fiatConnect: FiatConnectState
   swap: SwapState
   positions: PositionsState
-  appReview: AppReviewState
 }
 
 export interface PersistedRootState {
@@ -120,5 +117,4 @@ export interface PersistedRootState {
   localCurrency: LocalCurrencyState
   recipients: RecipientsState
   fiatExchanges: FiatExchangesState
-  appReview: AppReviewState
 }
