+++ src/redux/store.ts
@@ -138,7 +138,6 @@ export const setupStore = (initialState = {}, config = persistConfig) => {
           'imports',
           'paymentRequest',
           'fiatConnect',
-          'appReview',
         ],
       })
     )
